+++ src/main/java/domain/board/JumpingMovements.java
@@ -1,27 +1,26 @@
 package domain.board;
 
-import static domain.Path.DOWN_DOWN_DOWN_LEFT_LEFT_PATH;
-import static domain.Path.DOWN_DOWN_DOWN_RIGHT_RIGHT_PATH;
-import static domain.Path.DOWN_DOWN_LEFT_PATH;
-import static domain.Path.DOWN_DOWN_RIGHT_PATH;
-import static domain.Path.DOWN_PATH;
-import static domain.Path.LEFT_LEFT_DOWN_PATH;
-import static domain.Path.LEFT_LEFT_LEFT_DOWN_DOWN_PATH;
-import static domain.Path.LEFT_LEFT_LEFT_UP_UP_PATH;
-import static domain.Path.LEFT_LEFT_UP_PATH;
-import static domain.Path.LEFT_PATH;
-import static domain.Path.RIGHT_PATH;
-import static domain.Path.RIGHT_RIGHT_DOWN_PATH;
-import static domain.Path.RIGHT_RIGHT_RIGHT_DOWN_DOWN_PATH;
-import static domain.Path.RIGHT_RIGHT_RIGHT_UP_UP_PATH;
-import static domain.Path.RIGHT_RIGHT_UP_PATH;
-import static domain.Path.UP_PATH;
-import static domain.Path.UP_UP_LEFT_PATH;
-import static domain.Path.UP_UP_RIGHT_PATH;
-import static domain.Path.UP_UP_UP_LEFT_LEFT_PATH;
-import static domain.Path.UP_UP_UP_RIGHT_RIGHT_PATH;
+import static domain.board.Path.DOWN_DOWN_DOWN_LEFT_LEFT_PATH;
+import static domain.board.Path.DOWN_DOWN_DOWN_RIGHT_RIGHT_PATH;
+import static domain.board.Path.DOWN_DOWN_LEFT_PATH;
+import static domain.board.Path.DOWN_DOWN_RIGHT_PATH;
+import static domain.board.Path.DOWN_PATH;
+import static domain.board.Path.LEFT_LEFT_DOWN_PATH;
+import static domain.board.Path.LEFT_LEFT_LEFT_DOWN_DOWN_PATH;
+import static domain.board.Path.LEFT_LEFT_LEFT_UP_UP_PATH;
+import static domain.board.Path.LEFT_LEFT_UP_PATH;
+import static domain.board.Path.LEFT_PATH;
+import static domain.board.Path.RIGHT_PATH;
+import static domain.board.Path.RIGHT_RIGHT_DOWN_PATH;
+import static domain.board.Path.RIGHT_RIGHT_RIGHT_DOWN_DOWN_PATH;
+import static domain.board.Path.RIGHT_RIGHT_RIGHT_UP_UP_PATH;
+import static domain.board.Path.RIGHT_RIGHT_UP_PATH;
+import static domain.board.Path.UP_PATH;
+import static domain.board.Path.UP_UP_LEFT_PATH;
+import static domain.board.Path.UP_UP_RIGHT_PATH;
+import static domain.board.Path.UP_UP_UP_LEFT_LEFT_PATH;
+import static domain.board.Path.UP_UP_UP_RIGHT_RIGHT_PATH;
 
-import domain.Movement;
 import java.util.List;
 
 public enum JumpingMovements {
