+++ src/main/java/domain/board/Node.java
@@ -1,6 +1,5 @@
 package domain.board;
 
-import domain.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
