+++ src/main/java/util/ErrorHandler.java
@@ -1,4 +1,4 @@
-package domain.util;
+package util;
 
 import java.util.function.Supplier;
